# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pkm",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/search": {
      "get": {
        "summary": "Search",
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "exact",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Exact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/query": {
      "get": {
        "summary": "Run Query",
        "operationId": "run_query_api_query_get",
        "parameters": [
          {
            "name": "expr",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Expr"
            }
          },
          {
            "name": "expand",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Expand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/titles": {
      "get": {
        "summary": "Titles",
        "description": "Page-title completion for the editor's [[ / # popup.",
        "operationId": "titles_api_titles_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitlesPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/todos": {
      "get": {
        "summary": "Todos",
        "description": "Blocks whose text starts with a {{TODO}} marker, grouped by page.\nSQL narrows to TODO-containing candidates; the shared\npkm.todo matcher (the grammar's block-start rule, both bracket\nvariants, '> ' quote prefix) decides. Marker-based rather than\nrefs-based: the editor emits the bracket-less {{TODO}}, which\ncreates no ref row to query.",
        "operationId": "todos_api_todos_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupsPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/changed": {
      "get": {
        "summary": "Changed",
        "description": "Blocks touched in [since, until), grouped by page in the order\neach page was first touched. `since`/`until` are each\neither a 'YYYY-MM-DD' date (local midnight) or a full ISO datetime\n(naive = local time, aware = honoured as given); `until` defaults to\nnow and is exclusive. A block is 'new' when its created_at falls in\nthe window, else 'edited' -- `updated_at` is used as recorded, so a\nmove, indent, or rename also counts as an edit, and a block edited\nmore than once in the window shows only its latest edit time.",
        "operationId": "changed_api_changed_get",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Since"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangedPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/local/{path}": {
      "get": {
        "summary": "Get Local File",
        "operationId": "get_local_file_api_local__path__get",
        "parameters": [
          {
            "name": "path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlockGroup": {
        "properties": {
          "page_id": {
            "type": "integer",
            "title": "Page Id"
          },
          "page_title": {
            "type": "string",
            "title": "Page Title"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/GroupItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "page_id",
          "page_title",
          "items"
        ],
        "title": "BlockGroup"
      },
...
      "ChangedGroup": {
        "properties": {
          "page_id": {
            "type": "integer",
            "title": "Page Id"
          },
          "page_title": {
            "type": "string",
            "title": "Page Title"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChangedItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "page_id",
          "page_title",
          "items"
        ],
        "title": "ChangedGroup"
      },
      "ChangedItem": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "edited"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "text",
          "created_at",
          "updated_at",
          "status"
        ],
        "title": "ChangedItem"
      },
      "ChangedPayload": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/ChangedGroup"
            },
            "type": "array",
            "title": "Groups"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "since": {
            "type": "integer",
            "title": "Since"
          },
          "until": {
            "type": "integer",
            "title": "Until"
          }
        },
        "type": "object",
        "required": [
          "groups",
          "total",
          "since",
          "until"
        ],
        "title": "ChangedPayload",
        "description": "GET /api/changed: blocks whose updated_at falls in [since, until),\ngrouped by page in the order each page was first touched. `since`/\n`until` echo the resolved window (epoch ms) so a caller can see what\nwas actually queried, not just what it asked for."
      },
...
      "GroupItem": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "text"
        ],
        "title": "GroupItem"
      },
      "GroupsPayload": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/BlockGroup"
            },
            "type": "array",
            "title": "Groups"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "groups",
          "total"
        ],
        "title": "GroupsPayload",
        "description": "Shared by /api/unlinked and /api/todos."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QueryPayload": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/BlockGroup"
            },
            "type": "array",
            "title": "Groups"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "ref_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Ref Counts"
          }
        },
        "type": "object",
        "required": [
          "groups",
          "total",
          "ref_counts"
        ],
        "title": "QueryPayload",
        "description": "GET /api/query: groups plus per-operand match counts so an empty\nresult is steerable (bad query shape vs genuinely nothing)."
      },
...
      "SearchBlockHit": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "page_title": {
            "type": "string",
            "title": "Page Title"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "page_title",
          "snippet"
        ],
        "title": "SearchBlockHit"
      },
      "SearchPageHit": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "SearchPageHit"
      },
      "SearchPayload": {
        "properties": {
          "pages": {
            "items": {
              "$ref": "#/components/schemas/SearchPageHit"
            },
            "type": "array",
            "title": "Pages"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/SearchBlockHit"
            },
            "type": "array",
            "title": "Blocks"
          }
        },
        "type": "object",
        "required": [
          "pages",
          "blocks"
        ],
        "title": "SearchPayload"
      },
...
      "TitlesPayload": {
        "properties": {
          "titles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Titles"
          }
        },
        "type": "object",
        "required": [
          "titles"
        ],
        "title": "TitlesPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}